Added the option to rotate how the PSMove is held
This option is intended for when a PSNavi is attached to the PSMove
controller so that it can be done with the two triggers facing each
other. This changes the HMD alignment pose to bulb down and trigger
forward.

To enable this rotated mode set "psmove_rotate": true under
"psmove_settings" in steamvr.vrsettings.